requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
datapizza-ai==0.0.9
datapizza-ai-clients-openai==0.0.9
datapizza-ai-core==0.0.12
datapizza-ai-embedders-openai==0.0.5
datapizza-ai-vectorstores-qdrant==0.0.5
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
executing==2.2.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.22.0
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.5
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.0
referencing==0.37.0
rich==14.2.0
rpds-py==0.29.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
zipp==3.23.0
